requirements.txt
Django
djangorestframework
django-mptt
mozilla-django-oidc
python-dotenv
psycopg2-binary
africastalking
django-extensions
gunicorn
whitenoise